Prefix the * used for the select option with the table name.
This is needed when the generated query includes a join--we don't want to select all columns, just the columns on the primary table. This manifests itself when using a location query on a has_many :through association. In my app, this was causing my models to have the id of the joined record rather than the id of the primary record.